// src/enums/personaInformation.ts
export enum PersonaInformation {
    PERSONA_REQUIRED_NAME = "O nome da persona é obrigatório",
    PERSONA_REQUIRED_TEAM_ID = "O ID da equipe é obrigatório",
    PERSONA_REQUIRED_BRAND_ID = "O ID da marca é obrigatório",
    PERSONA_REQUIRED_GENDER = "O gênero da persona é obrigatório",
    PERSONA_REQUIRED_AGE = "A idade da persona é obrigatória e deve ser um número positivo",
    PERSONA_REQUIRED_LOCATION = "A localização da persona é obrigatória",
    PERSONA_REQUIRED_POSITION_DEGREE = "O cargo e formação acadêmica da persona são obrigatórios",
    PERSONA_REQUIRED_BELIEFS = "As crenças e interesses da persona são obrigatórios",
    PERSONA_REQUIRED_CONTENT_HABIT = "Os hábitos de consumo de conteúdo da persona são obrigatórios",
    PERSONA_REQUIRED_MAIN_OBJECTIVE = "O objetivo principal da persona é obrigatório",
    PERSONA_REQUIRED_CHALLENGE = "O desafio da persona é obrigatório",
    PERSONA_REQUIRED_FAVORITE_VOICE = "A voz favorita da persona é obrigatória",
    PERSONA_REQUIRED_BUYJOURNEY = "A jornada de compra da persona é obrigatória",
    PERSONA_REQUIRED_INTEREST_TRIGGER = "O gatilho de interesse da persona é obrigatório",
    PERSONA_NOT_FOUND = "Persona não encontrada",
    PERSONA_ALREADY_EXISTS = "Já existe uma persona com este nome para a marca informada",
    PERSONA_INVALID_ID = "ID da persona inválido",
    TEAM_ID_REQUIRED = "O ID da equipe é obrigatório",
    BRAND_ID_REQUIRED = "O ID da marca é obrigatório",
}